experimental commit : emoji in action

+++ src/main/java/edu/anadolu/RawTF.java
@@ -9,7 +9,7 @@ import org.apache.lucene.search.similarities.SimilarityBase;
 public class RawTF extends SimilarityBase {
 
     @Override
-    protected float score(BasicStats stats, float freq, float docLen) {
+    protected double score(BasicStats stats, double freq, double docLen) {
         return freq;
     }
